+++ Assets/LDtkUnity/Editor/Utility/HandleUtil.cs
@@ -41,6 +41,18 @@ namespace LDtkUnity.Editor
 
         public static void DrawText(string text, Vector3 pos, Vector2 guiOffset = default, Action onClicked = null)
         {
+            DrawText(text, pos, Color.black, guiOffset, onClicked);
+        }
+        
+        public static void DrawText(string text, Vector3 pos, Color color, Vector2 guiOffset = default, Action onClicked = null)
+        {
+            Vector3 guiPoint = HandleUtility.WorldToGUIPointWithDepth(pos);
+            //if camera is in front of the point, then don't draw it
+            if (guiPoint.z < 0)
+            {
+                return;
+            }
+            
             Handles.BeginGUI();
             
             GUIContent content = new GUIContent(text);
@@ -51,11 +63,17 @@ namespace LDtkUnity.Editor
 
             Rect textArea = HandleUtility.WorldPointToSizedRect(pos, content, style);
             
+
+#if UNITY_2019_3_OR_NEWER
+            const float yOffset = -3;
+#else
+            const float yOffset = -1;
+#endif
+            
             textArea.x += 1;
-            textArea.y -= 3;
+            textArea.y += yOffset;
             textArea.position += guiOffset;
-            //textArea.position += Vector2.one;
-                
+
             Rect backdropArea = new Rect(textArea);
             backdropArea.x += 1;
             backdropArea.y += 4;
@@ -66,13 +84,31 @@ namespace LDtkUnity.Editor
             {
                 onClicked?.Invoke();
             }
-                
-            Color color = new Color(0,0,0, 0.33f);
-                
-            EditorGUI.DrawRect(backdropArea, color);
+            
+            //maintain hue, maximise saturation, maintain value
+            Color.RGBToHSV(color, out float h, out float s, out float v);
+            Color backdropColor = Color.HSVToRGB(h, 1, v);
+            backdropColor.a = 0.75f;
+
+            Color textColor = GetTextColorForBackdrop(backdropColor);
+            style.normal = new GUIStyleState()
+            {
+                textColor = textColor
+            };
+            
+            EditorGUI.DrawRect(backdropArea, backdropColor);
+            
             GUI.Label(textArea, content, style);
             
             Handles.EndGUI();
+        }
+        
+        public static Color GetTextColorForBackdrop(Color backdropColor)
+        {
+            const float colorValue = 0.1f;
+            const float threshold = 0.8f;
+            Color.RGBToHSV(backdropColor, out float _, out float _, out float definitionValue);
+            return definitionValue > threshold ? new Color(colorValue, colorValue, colorValue) : Color.white;
         }
 
         public static void SelectIfNotAlreadySelected(GameObject obj)
@@ -82,6 +118,8 @@ namespace LDtkUnity.Editor
                 Selection.activeGameObject = obj;
             }
         }
+
+
     }
 }